feat: Extract control section
