+++ app/configurator/components/dataset-control-section.tsx
@@ -1,0 +1,137 @@
+import { t, Trans } from "@lingui/macro";
+import {
+  Box,
+  Button,
+  IconButton,
+  Link as MuiLink,
+  Typography,
+} from "@mui/material";
+import uniqBy from "lodash/uniqBy";
+import Link from "next/link";
+import { useMemo } from "react";
+
+import useDisclosure from "@/components/use-disclosure";
+import {
+  ControlSection,
+  ControlSectionContent,
+  SubsectionTitle,
+} from "@/configurator/components/chart-controls/section";
+import {
+  isConfiguring,
+  useConfiguratorState,
+} from "@/configurator/configurator-state";
+import { useDataCubesMetadataQuery } from "@/graphql/hooks";
+import SvgIcAdd from "@/icons/components/IcAdd";
+import SvgIcTrash from "@/icons/components/IcTrash";
+import { useLocale } from "@/locales/use-locale";
+
+import { DatasetDialog } from "./add-dataset-dialog";
+import { FiltersBadge } from "./chart-configurator";
+
+export const DatasetsControlSection = () => {
+  const [state, dispatch] = useConfiguratorState(isConfiguring);
+  const locale = useLocale();
+  const commonQueryVariables = {
+    sourceType: state.dataSource.type,
+    sourceUrl: state.dataSource.url,
+    locale,
+  };
+  const cubes = useMemo(() => {
+    const cubes = uniqBy(
+      state.chartConfigs.flatMap((x) => x.cubes),
+      (x) => x.iri
+    );
+    return cubes;
+  }, [state.chartConfigs]);
+  const [metadataQuery] = useDataCubesMetadataQuery({
+    variables: {
+      ...commonQueryVariables,
+      cubeFilters: cubes.map((cube) => ({
+        iri: cube.iri,
+      })),
+    },
+  });
+  const {
+    isOpen: isDatasetDialogOpen,
+    open: openDatasetDialog,
+    close: closeDatasetDialog,
+  } = useDisclosure();
+  return (
+    <ControlSection collapse>
+      <SubsectionTitle titleId="controls-data" gutterBottom={false}>
+        <Trans id="controls.section.datasets.title">Data Sources</Trans>{" "}
+        <FiltersBadge sx={{ ml: "auto", mr: 4 }} />
+      </SubsectionTitle>
+      <ControlSectionContent
+        aria-labelledby="controls-data"
+        data-testid="configurator-filters"
+      >
+        <Box
+          sx={{
+            gap: "0.5rem",
+            mb: "1rem",
+            display: "flex",
+            flexDirection: "column",
+          }}
+        >
+          {metadataQuery.data?.dataCubesMetadata.map((x) => {
+            return (
+              <Box
+                display="flex"
+                alignItems="center"
+                justifyContent="space-between"
+                key={x.iri}
+              >
+                <div>
+                  <Link href={`/browse?dataset=${x.iri}`} passHref>
+                    <MuiLink
+                      color="primary"
+                      underline="none"
+                      variant="caption"
+                      component="span"
+                    >
+                      Dataset
+                    </MuiLink>
+                  </Link>
+                  <br />
+                  <Typography variant="caption">{x.title}</Typography>
+                </div>
+                <div>
+                  {cubes.length > 1 ? (
+                    <IconButton
+                      onClick={() =>
+                        dispatch({
+                          type: "DATASET_REMOVE",
+                          value: { locale, iri: x.iri },
+                        })
+                      }
+                    >
+                      <SvgIcTrash />
+                    </IconButton>
+                  ) : null}
+                </div>
+              </Box>
+            );
+          })}
+        </Box>
+        <Box>
+          {cubes.length === 1 ? (
+            <Button
+              onClick={openDatasetDialog}
+              startIcon={<SvgIcAdd />}
+              variant="outlined"
+              size="small"
+            >
+              {t({ id: "chart.datasets.add", message: "Add dataset" })}
+            </Button>
+          ) : null}
+        </Box>
+        <DatasetDialog
+          state={state}
+          open={isDatasetDialogOpen}
+          onClose={closeDatasetDialog}
+        />
+      </ControlSectionContent>
+    </ControlSection>
+  );
+};
